pyproject.toml
[tool.poetry]
name = "garments-backend"
version = "0.1.0"
description = "Iot intehrated for basic shirts"
authors = ["sabbir, rakin"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.12"
uvicorn = {extras = ["standard"], version = "^0.34.1"}
sqlalchemy = "^2.0.40"
asyncpg = "^0.30.0"
pydantic = "^2.11.3"
psycopg2-binary = "^2.9.10"
alembic = "^1.15.2"
aio-pika = "^9.5.5"
paho-mqtt = "^2.1.0"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"
pytest = "^8.3.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
